NMproject: Script Based 'NONMEM' Model Development

Industrialisation of 'NONMEM' <https://www.iconplc.com/innovation/nonmem/> model development via end-to-end model development 'workflows'.

Package options

system_cmd

See ?system_cmd.

system_nm

See ?system_nm.

quiet_run

Sets the default value for the quiet argument in run_nm.

intern

Sets the default value for the intern argument in run_nm.

available_nm_types

A list of NONMEM subroutine names. This is modifiable in case future versions of NONMEM contain new subroutines.

nm_default_dirs

A optional (partially) named list of directories to be used by nm_create_analysis_project to create the project directory structure when making a new analysis projct. Required names are "models", "scripts" and "results". By default these will be set to "Models", "Scripts" and "Results", respectively. Additional characters (e.g. "SourceData") will ensure that those directories are presente in new projects.

kill_job

An option function to kill jobs before they are started. This is useful in infrastructures, where repeating a job conflicts with other jobs writing into the same directories. For example, in slurm grids this could be a function wrapper around a system("scancel ...") call.

nm.overwrite_behaviour

See ?overwrite_behaviour.

nm.force_render

Sets default value for the force argument in See nm_render.

nm.cmd_default

Sets default cmd field for all nm objects. See ?cmd for more information. The default value is "execute ctl_name -dir=run_dir". However in environments where a different command is needed, this is the place to change it.

nmtran_exe_path

Sets the value of the NMTRAN.exe program. This may be deprecated in future it is recommended to use nm_tran_command(). See ?nm_tran_command for more information.

code_library_path

Path to code library. By default this will point to the code library within the NMproject installation directory. See ?code_library for more information.

See also

Author

Maintainer: Tarj Sahota t.sahota0@gmail.com

Other contributors:

AstraZeneca [copyright holder]