Add/remove a covariate to a NONMEM model

Follows PsN coding conventions to add covariates into a model. The advantage is no need to create a .scm file, just directly modify the NONMEM control file contents. This function is used by covariate_step_tibble() for stepwise covariate model development.

add_cov(
  ctl,
  param,
  cov,
  state = 2,
  continuous = TRUE,
  time_varying,
  additional_state_text = list(),
  id_var = "ID",
  force = FALSE,
  force_TV_var = FALSE,
  init,
  lower,
  upper
)

remove_cov(
  ctl,
  param,
  cov,
  state = 2,
  continuous = TRUE,
  time_varying,
  id_var = "ID"
)

Arguments

ctl
An nm object or an object coercible to ctl_list.

param
Character. Name of parameter.

cov
Character. Name of covariate.

state
Numeric or character. Number or name of state (see details).

continuous
Logical (default = TRUE). Is covariate continuous?

time_varying
Optional logical. is the covariate time varying?

additional_state_text
Optional character (default = empty). Custom state variable to be passed to param_cov_text.

id_var
Character (default = "ID"). Needed if time_varying is missing.

force
Logical (default = `FALSE``). Force covariate in even if missing values found.

force_TV_var
Logical (default = FALSE). Force covariates only on TV notation parameters.

init
Optional numeric/character vector. Initial estimate of additional parameters.

lower
Optional numeric/character vector. lower bound of additional parameters.

upper
Optional numeric/character vector. Upper bound of additional parameters.

Details

Available states:

"2" or "linear"

PARCOV= ( 1 + THETA(1)*(COV -median))

"3" or "hockey-stick"

IF(COV.LE.median) PARCOV = ( 1 + THETA(1)*(COV - median)) IF(COV.GT.median) PARCOV = ( 1 + THETA(2)*(COV - median))

"4" or "exponential"

PARCOV= EXP(THETA(1)*(COV - median))

"5" or "power"

PARCOV= ((COV/median)**THETA(1))

"power1"

PARCOV= ((COV/median))

"power0.75"

PARCOV= ((COV/median)**0.75)

"6" or "log-linear"

PARCOV= ( 1 + THETA(1)*(LOG(COV) - log(median)))

remove_cov only works with covariates added with add_cov.

See also

Examples

if (FALSE) {

m1WT <- m1 %>% child("m1WT") %>%
  add_cov(param = "CL", cov = "WT", state = "power") %>%
  run_nm()

## compare results

rr(c(m1, m1WT))
summary_wide(c(m1, m1WT))
}

if (FALSE) {

m1noWT <- m1 %>% child("m1noWT") %>%
  remove_cov(param = "CL", cov = "WT") %>%
  run_nm()

## compare results

rr(c(m1, m1noWT))
summary_wide(c(m1, m1noWT))
}