Code Library

Function not designed for direct use. Instead use the RStudio code library addin on the Addins menu. This will open the shiny app. Select the file, and click "preview" to view and import() to bring into the "staging" area of your project. See vignette at https://tsahota.github.io/NMproject/ for a video showing use of the app. NONMEM control files will intentionally not be imported straight in the "Models" directory and instead go into "staging/Models". This staging location can be referred to when creating nm objects with new_nm(..., based_on = "staging/Models/[filename]")

code_library(
  extn = NULL,
  fields = "Description",
  viewer = TRUE,
  silent = FALSE,
  return_info = FALSE
)

Arguments

extn
Vector string of extensions to include (default = NULL includes all).

fields
Character vector of fields to extract.

viewer
Logical indicating if viewer should be used to display results (default=FALSE).

silent
Logical indicating if messages should be silenced (default=FALSE)

return_info
Logical (default = FALSE). Return data.frame of results (FALSE= returns file paths).

Details

requires getOption("code_library_path") to be set

See also