Extract parameter values

Pulls parameters, standard errors, OFVs and condition numbers out of ext files, applies transformations. This function is useful when numeric values are needed. rr is easier to read, however it returns characters. A wide and long format is available via two different functions.

coef_wide(m, trans = TRUE)

coef_long(m, trans = TRUE)

Arguments

m
An nm object.

trans
Logical (default = TRUE). Transform parameters using comments in $THETA/$OMEGA/$SIGMA.

Value

data.frame of extracted model parameter values. coef_wide() returns a data.frame in wide format. Vector valued objects m, will be stacked vertically with one row per run. coef_long() returns a data.frame in long format. Vector valued objects m, will be stacked horizontally.

NONMEM coding conventions used by NMproject

The convention for $THETA comments used by NMproject is value ; name ; unit ; transformation

e.g. $THETA 0.1 ; KA ; h-1 ; LOG

The options for THETA transformations are: LOG, LOGIT, RATIO and missing. LOG and LOGIT refer to log and logit transformed THETAs, respectively where the parameters should be back-transformed for reporting. RATIO refers to ratio data types, i.e. parameters that are positive and have a meaningful zero. Most parameters like KA, CL, EMAX fall into this category, but covariates effects which can go negative do not. RSEs are calculated for ratio data. Missing transformations are suitable for all other parameters, here no RSEs will be calculated, only raw SE values will be reported.

The convention for $OMEGA is similar but without a unit item: value ; name ; transformation

e.g. $OMEGA 0.1 ; IIV_KA ; LOG

The options for OMEGA are either LOG or missing. LOG indicating that the individual parameter distribution is log normally distributions and should be reported as a CV% (and associated RSE%) rather than as the raw NONMEM estimate.

The convention for $OMEGA is just : value ; name.

THETA transformations using trans = TRUE

The value of FINAL and RSE% (always accompanied with a % symbol in outputs) in the returned tibble is the reported standard error (where applicable) where \(\theta\) and \(se(\theta)\) are the NONMEM reported values of parameters and standard errors, respectively:

LOG

\(FINAL = exp(\theta), RSE = 100\sqrt(exp(se(\theta)^2) - 1)\)

RATIO

\(FINAL = \theta, RSE = 100se(\theta)/\theta\)

LOGIT

\(FINAL = 100/(1 + exp(-\theta)), SE = se(\theta)\)

missing

\(FINAL = \theta, SE = se(\theta)\)

OMEGA transformations using trans = TRUE

The value of FINAL and RSE% (always accompanied with a % symbol in outputs) in the returned tibble is the reported standard error (where applicable) where \(\omega^2\) and \(se(\omega^2)\) are the NONMEM reported values of parameters and standard errors, respectively

LOG

\(FINAL = 100\sqrt(exp(\omega^2) - 1), RSE = 100(se(\omega^2)/\omega^2)/2\)

missing

\(FINAL = \omega^2, SE = se(\omega^2)\)

SIGMA transformations using trans = TRUE

The value of FINAL and RSE% (always accompanied with a % symbol in outputs) in the returned tibble is the reported standard error (where applicable) where \(\sigma^2\) and \(se(\sigma^2)\) are the NONMEM reported values of parameters and standard errors, respectively. All sigmas are reported as standard deviations.

all sigmas

\(FINAL = \sqrt(\sigma^2), RSE = 100se(\sigma^2) / \sigma^2\)

See also