Get/set path to dataset

Mainly used to associate a dataset with an nm object. Requires ctl_contents to already be specified.

data_path(m, text)

Arguments

m
nm object

text
(optional) character. Path to input dataset

Value

if text is not specified, will return the data_path name otherwise will set data_path to the text provided

Examples

data_path,Get/set path to dataset,Get/set path to dataset — data_path,data_path.Rd,character(0),data_path,character(0),list(title = "Description", contents = "
Mainly used to associate a dataset with an nm object.\nRequires ctl_contents to already be specified.
"),list(description = "Mainly used to associate a dataset with an nm object.\nRequires ctl_contents to already be specified."),data_path(m, text),list(list(name = "m", description = "
nm object
"), list(name = "text", description = "
(optional) character. Path to input dataset
")),TRUE,
if (FALSE) {

## The following assumes a ctl file exists in the staging area:
##  staging/Models/ADVAN2.mod
## and a dataset
##  DerivedData/data.csv

m1 <- nm(run_id = "m1") %>%
      based_on("staging/Models/ADVAN2.mod") %>%
      data_path("DerivedData/data.csv")
      
data_path(m1)  ## display data name

}

,list(list(title = "Value", contents = "
if text is not specified, will return the data_path name\n otherwise will set data_path to the text provided
", slug = "value")),2021,list(name = "NMproject", version = "0.5.2.9000", authors = "Tarj Sahota"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "NMproject"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Get started\n
\n
\n  Reference\n
\n
\n  Changelog\n
", right = ""),list(.present = TRUE),list(version = "1.6.1"),FALSE