Make an OCC column for NONMEM IOV use

Creates and OCC column that increments in accordance to specified condition. To be used in a dplyr::mutate() statement dplyr::group_by()'d by "ID".

make_OCC_every_dose(d, dose_trigger, new_OCC_trigger)

Arguments

d
A data.frame. NONMEM ready input dataset.

dose_trigger
Logical expression for defining a dosing row.

new_OCC_trigger
Logical expression for defining when OCC should increment.

Examples

# create example object m1 from package demo files
exdir <- system.file("extdata", "examples", "theopp", package = "NMproject")
m1 <- new_nm(run_id = "m1", 
             based_on = file.path(exdir, "Models", "ADVAN2.mod"),
             data_path = file.path(exdir, "SourceData", "THEOPP.csv"))

d <- input_data(m1)

## OCC increments on every dosing interval with more than 4 samples
d %>% make_OCC_every_dose(!is.na(AMT), any(!is.na(DV)))

#>     ID  AMT  TIME    DV   WT OCC
#> 1    1 4.02  0.00    NA 79.6   1
#> 2    1   NA  0.00  0.74   NA   1
#> 3    1   NA  0.25  2.84   NA   1
#> 4    1   NA  0.57  6.57   NA   1
#> 5    1   NA  1.12 10.50   NA   1
#> 6    1   NA  2.02  9.66   NA   1
#> 7    1   NA  3.82  8.58   NA   1
#> 8    1   NA  5.10  8.36   NA   1
#> 9    1   NA  7.03  7.47   NA   1
#> 10   1   NA  9.05  6.89   NA   1
#> 11   1   NA 12.12  5.94   NA   1
#> 12   1   NA 24.37  3.28   NA   1
#> 49   2 4.40  0.00    NA 72.4   1
#> 50   2   NA  0.00  0.00   NA   1
#> 51   2   NA  0.27  1.72   NA   1
#> 52   2   NA  0.52  7.91   NA   1
#> 53   2   NA  1.00  8.31   NA   1
#> 54   2   NA  1.92  8.33   NA   1
#> 55   2   NA  3.50  6.85   NA   1
#> 56   2   NA  5.02  6.08   NA   1
#> 57   2   NA  7.03  5.40   NA   1
#> 58   2   NA  9.00  4.55   NA   1
#> 59   2   NA 12.00  3.01   NA   1
#> 60   2   NA 24.30  0.90   NA   1
#> 61   3 4.53  0.00    NA 70.5   1
#> 62   3   NA  0.00  0.00   NA   1
#> 63   3   NA  0.27  4.40   NA   1
#> 64   3   NA  0.58  6.90   NA   1
#> 65   3   NA  1.02  8.20   NA   1
#> 66   3   NA  2.02  7.80   NA   1
#> 67   3   NA  3.62  7.50   NA   1
#> 68   3   NA  5.08  6.20   NA   1
#> 69   3   NA  7.07  5.30   NA   1
#> 70   3   NA  9.00  4.90   NA   1
#> 71   3   NA 12.15  3.70   NA   1
#> 72   3   NA 24.17  1.05   NA   1
#> 73   4 4.40  0.00    NA 72.7   1
#> 74   4   NA  0.00  0.00   NA   1
#> 75   4   NA  0.35  1.89   NA   1
#> 76   4   NA  0.60  4.60   NA   1
#> 77   4   NA  1.07  8.60   NA   1
#> 78   4   NA  2.13  8.38   NA   1
#> 79   4   NA  3.50  7.54   NA   1
#> 80   4   NA  5.02  6.88   NA   1
#> 81   4   NA  7.02  5.78   NA   1
#> 82   4   NA  9.02  5.33   NA   1
#> 83   4   NA 11.98  4.19   NA   1
#> 84   4   NA 24.65  1.15   NA   1
#> 85   5 5.86  0.00    NA 54.6   1
#> 86   5   NA  0.00  0.00   NA   1
#> 87   5   NA  0.30  2.02   NA   1
#> 88   5   NA  0.52  5.63   NA   1
#> 89   5   NA  1.00 11.40   NA   1
#> 90   5   NA  2.02  9.33   NA   1
#> 91   5   NA  3.50  8.74   NA   1
#> 92   5   NA  5.02  7.56   NA   1
#> 93   5   NA  7.02  7.09   NA   1
#> 94   5   NA  9.10  5.90   NA   1
#> 95   5   NA 12.00  4.37   NA   1
#> 96   5   NA 24.35  1.57   NA   1
#> 97   6 4.00  0.00    NA 80.0   1
#> 98   6   NA  0.00  0.00   NA   1
#> 99   6   NA  0.27  1.29   NA   1
#> 100  6   NA  0.58  3.08   NA   1
#> 101  6   NA  1.15  6.44   NA   1
#> 102  6   NA  2.03  6.32   NA   1
#> 103  6   NA  3.57  5.53   NA   1
#> 104  6   NA  5.00  4.94   NA   1
#> 105  6   NA  7.00  4.02   NA   1
#> 106  6   NA  9.22  3.46   NA   1
#> 107  6   NA 12.10  2.78   NA   1
#> 108  6   NA 23.85  0.92   NA   1
#> 109  7 4.95  0.00    NA 64.6   1
#> 110  7   NA  0.00  0.15   NA   1
#> 111  7   NA  0.25  0.85   NA   1
#> 112  7   NA  0.50  2.35   NA   1
#> 113  7   NA  1.02  5.02   NA   1
#> 114  7   NA  2.02  6.58   NA   1
#> 115  7   NA  3.48  7.09   NA   1
#> 116  7   NA  5.00  6.66   NA   1
#> 117  7   NA  6.98  5.25   NA   1
#> 118  7   NA  9.00  4.39   NA   1
#> 119  7   NA 12.05  3.53   NA   1
#> 120  7   NA 24.22  1.15   NA   1
#> 121  8 4.53  0.00    NA 70.5   1
#> 122  8   NA  0.00  0.00   NA   1
#> 123  8   NA  0.25  3.05   NA   1
#> 124  8   NA  0.52  3.05   NA   1
#> 125  8   NA  0.98  7.31   NA   1
#> 126  8   NA  2.02  7.56   NA   1
#> 127  8   NA  3.53  6.59   NA   1
#> 128  8   NA  5.05  5.88   NA   1
#> 129  8   NA  7.15  4.73   NA   1
#> 130  8   NA  9.07  4.57   NA   1
#> 131  8   NA 12.10  3.00   NA   1
#> 132  8   NA 24.12  1.25   NA   1
#> 133  9 3.10  0.00    NA 86.4   1
#> 134  9   NA  0.00  0.00   NA   1
#> 135  9   NA  0.30  7.37   NA   1
#> 136  9   NA  0.63  9.03   NA   1
#> 137  9   NA  1.05  7.14   NA   1
#> 138  9   NA  2.02  6.33   NA   1
#> 139  9   NA  3.53  5.66   NA   1
#> 140  9   NA  5.02  5.67   NA   1
#> 141  9   NA  7.17  4.24   NA   1
#> 142  9   NA  8.80  4.11   NA   1
#> 143  9   NA 11.60  3.16   NA   1
#> 144  9   NA 24.43  1.12   NA   1
#> 13  10 5.50  0.00    NA 58.2   1
#> 14  10   NA  0.00  0.24   NA   1
#> 15  10   NA  0.37  2.89   NA   1
#> 16  10   NA  0.77  5.22   NA   1
#> 17  10   NA  1.02  6.41   NA   1
#> 18  10   NA  2.05  7.83   NA   1
#> 19  10   NA  3.55 10.21   NA   1
#> 20  10   NA  5.05  9.18   NA   1
#> 21  10   NA  7.08  8.02   NA   1
#> 22  10   NA  9.38  7.14   NA   1
#> 23  10   NA 12.10  5.68   NA   1
#> 24  10   NA 23.70  2.42   NA   1
#> 25  11 4.92  0.00    NA 65.0   1
#> 26  11   NA  0.00  0.00   NA   1
#> 27  11   NA  0.25  4.86   NA   1
#> 28  11   NA  0.50  7.24   NA   1
#> 29  11   NA  0.98  8.00   NA   1
#> 30  11   NA  1.98  6.81   NA   1
#> 31  11   NA  3.60  5.87   NA   1
#> 32  11   NA  5.02  5.22   NA   1
#> 33  11   NA  7.03  4.45   NA   1
#> 34  11   NA  9.03  3.62   NA   1
#> 35  11   NA 12.12  2.69   NA   1
#> 36  11   NA 24.08  0.86   NA   1
#> 37  12 5.30  0.00    NA 60.5   1
#> 38  12   NA  0.00  0.00   NA   1
#> 39  12   NA  0.25  1.25   NA   1
#> 40  12   NA  0.50  3.96   NA   1
#> 41  12   NA  1.00  7.82   NA   1
#> 42  12   NA  2.00  9.72   NA   1
#> 43  12   NA  3.52  9.75   NA   1
#> 44  12   NA  5.07  8.57   NA   1
#> 45  12   NA  7.07  6.59   NA   1
#> 46  12   NA  9.03  6.11   NA   1
#> 47  12   NA 12.05  4.57   NA   1
#> 48  12   NA 24.15  1.17   NA   1