Create a new (parent) nm object

escription

new_nm(based_on, run_id = NA_character_, data_path, cmd)

Arguments

based_on
Character. Prior ctl file to base run on

run_id
Character. Run identifier

data_path
Character. Path to dataset

cmd
Optional character. PsN command to use. If unspecified will use getOption("nm.cmd_default"). Use glue notation for inheritance. See details.

Value

An object of class nm_list. Attributes

Details

Create a new parent nm object. Normally the first NONMEM you create will be using this functions. Subsequent objects created with the child() function will inherit the properties of the parent run.

The cmd field uses glue notation. So instead of specifying execute runm1.mod -dir=m1, it is best to specify execute {ctl_name} -dir={run_dir}. The values of ctl_name and run_dir refer to object fields and if these change value like when the child() function is used to create a separate child object, the cmd field will update automatically.

object fields

Each field has a corresponding function (documented in nm_getsetters) of the same name to access and modify it's value.

type

The PsN run type. Default is execute.

run_id

The run identifier. E.g. m1.

run_in

The directory to copy control files and run NONMEM. Default = "Models".

executed

For internal use.

ctl_contents

Stores the contents of the control file to be written to disk when the run_nm() function is used on the object.

data_path

Path to the NONMEM ready dataset (from base project directory).

cmd

See details above.

cores

Numbers of cores to use. Requires a cmd value that uses the {cores} glue field.

run_dir

PsN directory to run the NONMEM run. Default is to the be the same as the run_id for simplicity.

results_dir

Location to store results files from diagnostic reports executed with nm_render().

unique_id

For internal use.

lst_path

Normally does not require setting. Path to the expected .lst file.

See also

Examples

if (FALSE) {

m1 <- new_nm(run_id = "m1",
            based_on = "staging/Models/run1.mod",
            data_path = "DerivedData/data.csv",
            cmd = "execute -run_on_sge -sge_prepend_flags='-V' {ctl_name} -dir={run_dir}") %>%
     fill_input() %>%
     run_nm()

m1 ## display object fields
cmd(m1)
ctl_name(m1)
run_dir(m1)

}