Create analysis project

Create analysis project

nm_create_analysis_project(
  path,
  dirs = nm_default_dirs(),
  style = c("simple", "starters"),
  use_renv = TRUE,
  ...
)

Arguments

path
character path (relative or absolute) to project. If just specifying a name, this will create the analysis project in the current working directory

dirs
character list or vector. Default = `nm_default_dirs()`

style
character. Either "simple" (default) or "starters". See details

use_renv
logical (default = TRUE). should renv be used or not in project

...
arguments passted to `starters::create_analysis_package`

Details

if `style = "starters"` the function works as light wrapper around `starters::create_analysis_project`. There will be restriction that `name` will not be able to contain underscores and will need to be comprised of only letters, numbers and periods. if `style = "simple"` the function sacrifices some of the R package-like structure so that underscores can be used.