functions to access and modify fields of nm objects

functions to access and modify fields of nm objects

run_dir(m, text)

cmd(m, text)

type(m, text)

parent_run_id(m, text)

parent_run_in(m, text)

parent_ctl_name(m, text)

parent_results_dir(m, text)

unique_id(m, text)

ctl_name(m, text)

results_dir(m, text)

cores(m, text)

parafile(m, text)

executed(m, text)

lst_path(m, text)

run_in(m, text)

run_id(m, text)

result_files(m, text)

Arguments

m
nm object

text
optional character for replacing field. If present function will modify field (of same name as function) otherwise will return value of field (of same name as function)

Details

Easiest way to see all fields of an object is to type the object into the console and hit enter. This will display the value of each field. some fields like cmd are glue fields. In these cases inserting expressions inside braces in text will evaluate the expression

Examples

nm_getsetters,functions to access and modify fields of nm objects,functions to access and modify fields of nm objects — nm_getsetters,Source: R/nm_gettersetters.R,nm_getsetters.Rd,character(0),c("nm_getsetters", "run_dir", "cmd", "type", "parent_run_id", "parent_run_in", "parent_ctl_name", "parent_results_dir", "unique_id", "ctl_name", "results_dir", "cores", "parafile", "executed", "lst_path", "run_in", "run_id", "result_files"),character(0),list(title = "Description", contents = "
functions to access and modify fields of nm objects
"),list(description = "functions to access and modify fields of nm objects"),run_dir(m, text)

cmd(m, text)

type(m, text)

parent_run_id(m, text)

parent_run_in(m, text)

parent_ctl_name(m, text)

parent_results_dir(m, text)

unique_id(m, text)

ctl_name(m, text)

results_dir(m, text)

cores(m, text)

parafile(m, text)

executed(m, text)

lst_path(m, text)

run_in(m, text)

run_id(m, text)

result_files(m, text),list(list(name = "m", description = "
nm object
"), list(name = "text", description = "
optional character for replacing field. If present function will\nmodify field (of same name as function) otherwise will return value of\nfield (of same name as function)
")),TRUE,
if (FALSE) {

## set cmd field of m1
m1 <- m1 %>% cmd("execute {ctl_name} -dir={run_dir}")

m1 %>% cmd() ## display command
## can also view field when viewing object
m1

}

,list(list(title = "Details", contents = "
Easiest way to see all fields of an object is to type the object\n  into the console and hit enter. This will display the value of each field.\n  some fields like cmd are glue fields.  In these cases inserting\n  expressions inside braces in text will evaluate the expression
", slug = "details")),2021,list(name = "NMproject", version = "0.5.9.9000", authors = "Tarj Sahota"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "NMproject"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Get started\n
\n
\n  Reference\n
\n
\n  Changelog\n
", right = "
\n  \n    \n     \n  \n
"),list(.present = TRUE),list(version = "1.6.1"),FALSE