Execution related functions to access and modify fields of nm objects

The fields of an object can be viewed by printing the object. Each field has a corresponding function of the same name to access and modify it's value.

cores(m, text)

parafile(m, text)

walltime(m, text)

executed(m, text)

Arguments

m
An nm object.

text
Optional character for replacing field. If present function will modify field (of same name as function) otherwise will return value of field (of same name as function).

Details

Easiest way to see all fields of an object is to type the object into the console and hit enter. This will display the value of each field. some fields like cmd are glue fields. In these cases inserting expressions inside braces in text will evaluate the expression

The fundamental structure of all these functions is the same:

To access the value of a field: m %>% fieldname() or equivalently fieldname(m)

To modify the value of a field: m <- m %>% fieldname("newvalue")