render function for nm objects

render function for nm objects

nm_render(
  m,
  input,
  output_file = NA,
  args = list(),
  force = getOption("nm.force_render"),
  async = FALSE,
  ...
)

Arguments

m
nm object

input
character. Same as rmarkdown::render() arg

output_file
character. Same as rmarkdown::render() arg

args
list. Same as "params" arg in rmarkdown::render()

force
logical (default = getOption("nm.force_render")). will force execution

async
experimental option to use future package

...
additional argument passed to rmarkdown::render()

Details

input must refer to a properly specified Rmd document. The R markdown template "model diagnostic" in RStudio sets this up for you.

These R markdown templates are usable as R Notebooks (e.g. for code development and debugging) if the object .m is defined in the global work space first.

Examples

if (FALSE) {
  m1 %>% nm_render("Scripts/basic_gof.Rmd")

  ## to run "Scripts/basic_gof.Rmd" as an R Notebook
  ## first define .m
  
  .m <- m1 ## Now you can run "Scripts/basic_gof.Rmd" as a Notebook
}