Generate a summary of NONMEM results

Get wide (or a long) tibble showing summary results.

summary_wide(
  ...,
  include_fields = character(),
  parameters = c("none", "new", "all"),
  m = TRUE,
  trans = TRUE
)

summary_long(..., parameters = c("none", "new", "all"))

Arguments

...
Arguments passed to summary(), usually a vector of nm object + options.

include_fields
Character vector of nm object fields to include as columns in the output. Default is empty.

parameters
Character. Either "none" (default), "new", or "all" indicating whether parameter values should be included in the summary tibble. Specifying "new" means that only parameters that aren't in the parent run are included in outputs. This is useful if wanting to know the value of an added parameter but not all the parameters (e.g. in a covariate analysis).

m
Logical (default = TRUE). Should model object be included as the m column.

trans
Logical (default = TRUE). Should parameters be transformed in accordance with $THETA/$OMEGA/$SIGMA comments. This is only valid if parameters is "new" or "all.