Get/set nm_tran_command

The function nm_tran() needs the location of NMTRAN.exe to function. This is guessed at package load, assuming PsN is on the $PATH environmental variable. If this is not the case, then you can manually set the path and command used.

nm_tran_command(text)

Arguments

text
Optional character. If specified will set nm_tran_command otherwise it will display the current option value.

Value

If text is missing will get and return the current NMTRAN command.

Details

text can just be the path to NMTRAN.exe in which case nm_tran_command will use the format /path/to/NMTRAN.exe < {ctl_name} to launch NMTRAN.exe where {ctl_name} is the name of the control file. Specifying

nm_tran_command("/path/to/NMTRAN.exe < {ctl_name}") is equivalent to: nm_tran_command("/path/to/NMTRAN.exe")

More complicated formats are possible with different installations which can be seen examples.

As with all NMproject configuration options set this up either at the beginning of your script, in your .Rprofile or for all users in Rprofile.site. See FAQ for setting up configuration options permanently.

See also

Examples

if (FALSE) {

# the following two are equivalent
nm_tran_command("/opt/NONMEM/nm75/tr/NMTRAN.exe")
nm_tran_command("/opt/NONMEM/nm75/tr/NMTRAN.exe < {ctl_name}")
}