Run NONMEM jobs

Run nm objects. Uses system_nm() to submit the cmd() value of object.

run_nm(
  m,
  ignore.stdout = TRUE,
  ignore.stderr = TRUE,
  quiet = getOption("quiet_run"),
  intern = getOption("intern"),
  force = FALSE,
  cache_ignore_cmd = FALSE,
  cache_ignore_ctl = FALSE,
  cache_ignore_data = FALSE
)

run_nm_batch(m, threads = 10, ...)

Arguments

m

An nm object.

ignore.stdout

Logical (default=TRUE). Parameter passed to system_nm().

ignore.stderr

Logical (default=TRUE). Parameter passed to system_nm().

quiet

Logical (default=FALSE). Should system_nm() output be piped to screen?

intern

Logical. intern argument to be passed to system_nm().

force

Logical (default = FALSE). Force run even results unchanged.

cache_ignore_cmd

Logical (default = FALSE). Should check cmd field with cache?

cache_ignore_ctl

Logical (default = FALSE). Should check control file contents with cache?

cache_ignore_data

Logical (default = FALSE). Should check dataset with cache?

threads

Numeric. Number of threads to run concurrently.

...

Additional arguments passed to run_nm().

Value

m with job_info fields populated.

Details

In grid environment it is recommended to run nm_tran() via the RStudio 'Addin' prior to executing this code.

By default, when highlighting code and evaluating it via an RStudio app, run_nm() will not execute and will just return the nm object.

For vector nm objects of length more than 1, all runs will be launched at the same time. This could overwhelm resources if not in a grid environment. In this case see run_nm_batch() for batched execution of a vector valued nm object.

run_nm_batch is a variant of run_nm() containing a threads argument that will submit run_nm()'s in batches and wait for them to complete. If you need all the runs to complete ensure you use a wait_finish() statement afterwards as R console will only be blocked for until the last batch has been submitted which will be before all runs have completed

See also

Examples

## requires NONMEM to be installed
if (FALSE) {
m1 <- new_nm(
  run_id = "m1",
  based_on = "staging/Models/ADVAN2.mod",
  data_path = "DerivedData/data.csv"
) %>%
  cmd("execute {ctl_name} -dir={run_dir}") %>%
  fill_input() %>%
  run_nm()
}