Remove temporary NONMEM files

NONMEM produces a lot of temporary files which can add up to a lot of disk space. One strategy to remove this is to use the clean option in the PsN command. However, this can automatically remove files as soon as the run finishes that may be useful for debugging. ls_tempfiles()`` allows you to list the paths of all temporary files, for a single run or for all runs for inspection and deletion. clean_run()is a wrapper function that runsls_tempfiles()and deletes everything returned. For safety is limited to only deleting files associated withnm` objects though.

ls_tempfiles(
  object = ".",
  output_loc = c("run_dir", "base"),
  run_files = NA_character_,
  include_slurm_files = TRUE,
  ctl_extension = "mod",
  include_psn_exports = FALSE
)

clean_run(m, output_loc = c("run_dir", "base"), include_slurm_files = TRUE)

Arguments

object
Either an nm object or path to project (default = "."). If a path is specified, the function will look for all runs in the directory (including subdirectories).

output_loc
Optional character for locating files. Either "run_dir" (default) for PsN execution or "base" for "nmfe" execution.

run_files
Optional character vector. Search amongst only these files instead. Default value NA searches based on object.

include_slurm_files
Logical (default = TRUE). Include files generated by Slurm.

ctl_extension
Character. Extension of control file (default = "mod")

include_psn_exports
Logical (default = FALSE). Considers files that PsN exports to the run_in directory as temporary

m
An nm object

Details

Setting include_psn_exports = TRUE will break 'Pirana' and 'xpose' capability as these software use exported files.

Examples

if (FALSE) {

ls_tempfiles(m1) ## display all temp files associated with m1

m1 %>%
  ls_tempfiles() %>%
  unlink() ## delete all m1 temp files

## above line is equivalent to:
clean_run(m1)

ls_tempfiles() ## display all temp files in analysis project

ls_tempfiles() %>% unlink() ## remove all temp files in analysis project
}