NMproject: Script Based 'NONMEM' Model Development

Industrialisation of 'NONMEM' <https://www.iconplc.com/innovation/nonmem/> via fully and rapidly reusable model development 'workflows' entirely within 'RStudio'. Quickly get started with new models by importing 'NONMEM' templates from the built-in code library. Manipulate 'NONMEM' code from within R either via the tracked 'manual edit' interface or 'programmatically' via convenience functions. Script 'workflows' by piping sequences of model building steps from control file creation, to execution, to post-processing and evaluation. Run caching makes 'workflows' R markdown friendly for easy documentation of thoughts and modelling decisions alongside executable code. Share, reuse and recycle 'workflows' for new problems.

Package options

See options() for information on how to modify these:

system_cmd

See ?system_cmd.

system_nm

See ?system_nm.

quiet_run

Sets the default value for the quiet argument in run_nm.

intern

Sets the default value for the intern argument in run_nm.

available_nm_types

A list of NONMEM subroutine names. This is modifiable in case future versions of NONMEM contain new subroutines.

nm_default_dirs

An optional (partially) named list of directories to be used by nm_create_analysis_project to create the project directory structure when making a new analysis project. Required names are "models", "scripts" and "results". By default these will be set to "Models", "Scripts" and "Results", respectively. Additional characters (e.g. "SourceData") correspond to additional modelling directories to be created. See ?nm_default_fields for more information.

kill_job

An optional function to kill jobs before they are started. This is useful in infrastructures, where repeating a job conflicts with other jobs writing into the same directories. For example, in Slurm grids this could be a function wrapper around a system("scancel ...") call.

nm.overwrite_behaviour

See ?overwrite_behaviour.

nm.force_render

Sets default value for the force argument in See nm_render.

nm_default_fields

Sets default fields for all nm objects. See ?nm_default_fields for more information.

nmtran_exe_path

Sets the value of the NMTRAN.exe program. This may be deprecated in future it is recommended to use nm_tran_command(). See ?nm_tran_command for more information.

code_library_path

Path to code library. By default this will point to the code library within the NMproject installation directory. See ?code_library for more information.

See also

Author

Maintainer: Tarj Sahota t.sahota0@gmail.com [copyright holder]

Other contributors:

AstraZeneca [copyright holder]

Nuria Buil Bruna nuria.buil.bruna@gmail.com [contributor]

Stein Schalkwijk steinschalkwijk@gmail.com [contributor]