Exclude rows of NONMEM dataset

A mechanism for excluding outliers during data cleaning. Create exploratory plots, identify rows of the dataset to be considered outliers for exclusion, and then feed that filtered dataset into this function to exclude them from the dataset. Requires a corresponding IGNORE statement - see argument descriptions for more details.

exclude_rows(d, dexcl, exclude_col = "EXCL")

Arguments

d
A data.frame for containing the full NONMEM dataset. Should contain a column for identifying excluded rows named with the exclude_col argument.

dexcl
A smaller data.frame consisting of rows to be ignored. Need not contain all columns of d but each column should be present in d.

exclude_col
Character (default = "EXCL"). Name of a binary exclude column in d. This should be accompanied with a IGNORE=(EXCL.GT.0) statement in $DATA.

Value

A modified version of d with exclude_col set to 1 for rows coinciding with dexcl.

See also

Examples

if (FALSE) {
## use with dplyr
dexcl <- d %>%
  filter(ID == 23, TIME > 18, TIME < 24) %>%
  select(ID, TIME, DV, EXCL)
dexcl ## view rows to be excluded
d <- d %>% exclude_rows(dexcl)
}